packages/lm-ocr/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lm-ocr"
version = "0.1.0"
description = "LedgerMate - Receipt OCR (Upstage Document Digitization)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0,<3",
  "pydantic>=2.7.0,<3",
  "python-dateutil>=2.8.2,<3",
  "python-dotenv>=1.0.0,<2",
]

[project.scripts]
lm-ocr = "lm_ocr.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lm_ocr*"]
